pyproject.toml
[project]
name = "tjr104-t01"
version = "0.1.0"
description = ""
authors = [
    {name = "Carl",email = "eszaqw6207@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "folium (>=0.20.0,<0.21.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "pandas (<3.0.0)",
    "streamlit (>=1.53.1,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "selenium (>=4.40.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "geopandas (>=1.1.2,<2.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "googlemaps (>=4.10.0,<5.0.0)",
    "streamlit-folium (>=0.26.1,<0.27.0)",
    "redis (>=7.1.1,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
